Move "getNPC" that was introduced in commit https://github.com/edubart/otclient/commit/00729bbc2ea8e359d9ab626db8a1154055110f02 from Spawn to CreatureType
I didn't have a closer look at how his code was structured, what he
basically did
is that he set all creatures in a spawn as NPC's even if it's a monster
which is
so erroneous.

Highlights:
- Add branch prediction macros
- Minor code style fixes & some others

Hopefully the branch prediction thing will speed up OTB since it's
awfully slow.